[api]: Improve Star/DSO Catalog Generator

+++ api/src/test/kotlin/DsoIdFetcher.kt
@@ -3,10 +3,10 @@ import nebulosa.simbad.SimbadService
 
 data class DsoIdFetcher(
     private val catalogType: String,
-    private val lastId: Int = 0,
-    private val maxId: Int = -1,
+    private val lastId: Long = 0L,
+    private val maxId: Long = -1L,
     private val limit: Int = 200,
-) : Fetcher<IntArray> {
+) : Fetcher<LongArray> {
 
     private val builder = QueryBuilder()
 
@@ -22,9 +22,9 @@ data class DsoIdFetcher(
         builder.add(SortBy(oidRef))
     }
 
-    override fun fetch(service: SimbadService): IntArray {
-        val rows = service.query(builder.build()).execute().body() ?: return IntArray(0)
-        return IntArray(rows.size) { rows[it].getField("oidref").toInt() }
+    override fun fetch(service: SimbadService): LongArray {
+        val rows = service.query(builder.build()).execute().body() ?: return LongArray(0)
+        return LongArray(rows.size) { rows[it].getField("oidref").toLong() }
     }
 
     companion object {
