+++ api/src/test/kotlin/StarIdFetcher.kt
@@ -2,10 +2,10 @@ import nebulosa.adql.*
 import nebulosa.simbad.SimbadService
 
 data class StarIdFetcher(
-    private val lastId: Int = 0,
-    private val maxId: Int = -1,
+    private val lastId: Long = 0L,
+    private val maxId: Long = -1L,
     private val limit: Int = 200,
-) : Fetcher<IntArray> {
+) : Fetcher<LongArray> {
 
     private val builder = QueryBuilder()
 
@@ -29,9 +29,9 @@ data class StarIdFetcher(
         builder.add(SortBy(oidRef))
     }
 
-    override fun fetch(service: SimbadService): IntArray {
-        val rows = service.query(builder.build()).execute().body() ?: return IntArray(0)
-        return IntArray(rows.size) { rows[it].getField("oidref").toInt() }
+    override fun fetch(service: SimbadService): LongArray {
+        val rows = service.query(builder.build()).execute().body() ?: return LongArray(0)
+        return LongArray(rows.size) { rows[it].getField("oidref").toLong() }
     }
 
     companion object {
